+++ src/App/AppContext.ts
@@ -1,0 +1,19 @@
+import { createContext } from 'react';
+
+export interface AppProps {
+  user?: User;
+  scope: string;
+  packages: any[];
+}
+
+export interface User {
+  username: string;
+}
+
+export interface AppContextProps extends AppProps {
+  setUser: (user?: User) => void;
+}
+
+const AppContext = createContext<undefined | AppContextProps>(undefined);
+
+export default AppContext;