+++ src/components/AvatarTooltip/index.ts
@@ -1,4 +1,0 @@
-import { AvatarTooltip } from './AvatarTooltip';
-
-export { AvatarTooltip };
-export default AvatarTooltip;