+++ src/components/Header/HeaderLogo.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import styled from '@emotion/styled';
 
 import Logo from '../Logo';
 
@@ -8,7 +9,14 @@ interface Props {
 
 const HeaderLogo: React.FC<Props> = ({ logo }) => {
   if (logo) {
-    return <img alt="logo" height="40px" src={logo} />;
+    const Wrapper = styled('div')({
+      fontSize: 0,
+    });
+    return (
+      <Wrapper>
+        <img alt="logo" height="40px" src={logo} />
+      </Wrapper>
+    );
   }
 
   return <Logo />;