+++ src/components/Package/Package.tsx
@@ -9,7 +9,7 @@ import fileSizeSI from '../../utils/file-size';
 import { formatDate, formatDateDistance } from '../../utils/package';
 import Tooltip from '../../muiComponents/Tooltip';
 import { isURL } from '../../utils/url';
-import { downloadHandler } from '../ActionBar/ActionBar';
+import { downloadTarball } from '../ActionBar';
 import ListItem from '../../muiComponents/ListItem';
 import Grid from '../../muiComponents/Grid';
 
@@ -104,7 +104,7 @@ const Package: React.FC<PackageInterface> = ({
       <OverviewItem>
         <Icon name="time" />
         <Published>{`Published on ${formatDate(time)} •`}</Published>
-        {`${formatDateDistance(time)} ago`}
+        {formatDateDistance(time)}
       </OverviewItem>
     );
 
@@ -114,7 +114,6 @@ const Package: React.FC<PackageInterface> = ({
       <a href={homepage} target={'_blank'}>
         <Tooltip aria-label={'Homepage'} title={'Visit homepage'}>
           <IconButton aria-label={'Homepage'}>
-            {/* eslint-disable-next-line react/jsx-max-depth */}
             <HomeIcon />
           </IconButton>
         </Tooltip>
@@ -128,7 +127,6 @@ const Package: React.FC<PackageInterface> = ({
       <a href={bugs.url} target={'_blank'}>
         <Tooltip aria-label={'Bugs'} title={'Open an issue'}>
           <IconButton aria-label={'Bugs'}>
-            {/* eslint-disable-next-line react/jsx-max-depth */}
             <BugReport />
           </IconButton>
         </Tooltip>
@@ -140,10 +138,9 @@ const Package: React.FC<PackageInterface> = ({
     dist.tarball &&
     isURL(dist.tarball) && (
       // eslint-disable-next-line
-      <a onClick={() => downloadHandler(dist.tarball.replace(`https://registry.npmjs.org/`, window.location.href))} target={'_blank'}>
+      <a onClick={downloadTarball(dist.tarball.replace(`https://registry.npmjs.org/`, window.location.href))} target={'_blank'}>
         <Tooltip aria-label={'Download the tar file'} title={'Download tarball'}>
           <IconButton aria-label={'Download'}>
-            {/* eslint-disable-next-line react/jsx-max-depth */}
             <DownloadIcon />
           </IconButton>
         </Tooltip>
@@ -155,7 +152,6 @@ const Package: React.FC<PackageInterface> = ({
       <Grid container={true} item={true} xs={12}>
         <Grid item={true} xs={true}>
           <WrapperLink to={`/-/web/detail/${packageName}`}>
-            {/* eslint-disable-next-line react/jsx-max-depth */}
             <PackageTitle className="package-title">{packageName}</PackageTitle>
           </WrapperLink>
         </Grid>